+++ src/colors/syntax/ts.colors.ts
@@ -12,6 +12,48 @@ const tsColors: SyntaxColors = (tokens) => {
     },
 
     {
+      name: 'Typescript: Language classes',
+      scope: ['support.constant.math.ts'],
+      settings: tokens.class.library
+    },
+
+    /**
+     * Interfaces
+     */
+    {
+      name: 'Typescript: Extended interface',
+      scope: 'meta.interface.ts entity.other.inherited-class.ts',
+      settings: tokens.type.custom
+    },
+
+    /**
+     * Enums
+     */
+    {
+      name: 'Typescript: Enum member',
+      scope: 'variable.other.enummember.ts',
+      settings: tokens.enum.member
+    },
+
+    {
+      name: 'Typescript: Type punctuation',
+      scope: [
+        'meta.type.tuple.ts meta.brace.square.ts',
+        'punctuation.definition.typeparameters.begin.ts',
+        'punctuation.definition.typeparameters.end.ts'
+      ],
+      settings: tokens.type.punctuation
+    },
+
+    {
+      name: 'Typescript: Template expression punctuation',
+      scope: [
+        'punctuation.definition.template-expression.begin.ts',
+        'punctuation.definition.template-expression.end.ts'
+      ],
+      settings: tokens.puntuaction.embedded
+    },
+    {
       name: 'Arrow function. =>',
       scope: 'storage.type.function.arrow.ts',
       settings: tokens.javascript.arrowFunction
@@ -22,7 +64,10 @@ const tsColors: SyntaxColors = (tokens) => {
         'keyword.operator.new.ts',
         'keyword.operator.in.ts',
         'keyword.operator.of.ts',
-        'keyword.operator.typeof.ts'
+        'keyword.operator.typeof.ts',
+        'keyword.operator.expression.instanceof.ts',
+        'keyword.operator.expression.in.ts',
+        'keyword.operator.expression.keyof.ts'
       ],
       settings: tokens.keyword.other
     }